@@ -23,34 +23,43 @@
     "bin": {
         "example": "./bin/example"
     },
+    "files": [
+        "bin",
+        "lib"
+    ],
     "scripts": {
         "reinstall": "rimraf ./package-lock.json ./node_modules && npm i",
-        "postinstall": "npm run compile",
+        "postinstall": "npm run bundle",
         "precompile:dev": "npm run lint",
         "compile:dev": "npm run compile -- -p tsconfig.dev.json",
         "precompile": "rimraf ./lib",
         "compile": "tsc",
         "compile:w": "npm run compile -- -w",
         "lint": "eslint . --ext .ts",
+        "bundle": "npx rollup -c",
         "preexample": "npm run lint && npm run compile",
         "example": "node ./bin/example"
     },
     "dependencies": {
-        "yargs": "~13.3.0"
+        "yargs": "~15.3.1"
     },
     "devDependencies": {
-        "@types/node": "8",
-        "@types/yargs": "~13.0.0",
+        "@types/node": "10",
+        "@types/yargs": "~15.0.4",
         "@typescript-eslint/eslint-plugin": "^2.27.0",
         "@typescript-eslint/parser": "^2.27.0",
+        "@rollup/plugin-typescript": "3.1.1",
+        "builtin-modules": "^3.1.0",
         "eslint": "^6.8.0",
         "eslint-config-prettier": "^6.10.1",
         "eslint-plugin-import": "^2.20.2",
         "eslint-plugin-prettier": "^3.1.2",
-        "husky": "^4.2.3",
-        "lint-staged": "^10.1.2",
+        "husky": "^4.2.5",
+        "lint-staged": "^10.1.3",
         "prettier": "^2.0.4",
-        "rimraf": "^2.6.3",
-        "typescript": "^3.5.3"
+        "rimraf": "^3.0.2",
+        "rollup": "^2.4.0",
+        "rollup-plugin-terser": "^5.3.0",
+        "typescript": "^3.8.3"
     }
 }
